Fix compilation of C++ test cases.
  Also, added copyright notice an test case files.
  
  modified:
    src.c++/examples/rank.cpp
    src.c++/test/test_denserow.cpp
    src.c++/test/test_rank_double.cpp
    src.c++/test/test_rank_int.cpp
    src.c++/test/test_sparserow.cpp

Original commit message in repo xenia:rheinfall (rev 166)
   revno: 166
   committer: Riccardo Murri <riccardo.murri@gmail.com>
   branch nick: rheinfall
   timestamp: Sun 2010-11-21 17:42:16 +0100

+++ trunk/src.c++/test/test_rank_double.cpp
@@ -1,5 +1,35 @@
+/**
+ * @file   test_rank_long.cpp
+ *
+ * Correctness test for the `Rheinfall` C++ implementation, with `long
+ * double` coefficients.
+ *
+ * @author  riccardo.murri@gmail.com
+ * @version $Revision$
+ */
+/*
+ * Copyright (c) 2010 riccardo.murri@gmail.com.  All rights reserved.
+ *
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, write to the Free Software
+ * Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301 USA
+ *
+ */
 
-typedef double val_t;
+#include <config.h>
+
+typedef long double val_t;
 typedef long coord_t;
 
 #include "test_rank.cpp.inc"
